pyproject.toml
[project]
name = "x-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.45.0",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "tweepy>=4.16.0",
    "typer[all]>=0.17.4",
]
